refactoring: split NexusRepositoryIndexerImpl
 - moves about half of the code into a dedicated class
 - queries moved to NexusRepositoryQueries
 - index management remains in NexusRepositoryIndexManager
